chore(material): lazy import editors
